use private profile directory for Chrome
It's not possible to force launching a new instance
of Chrome via the command line, but it's required to
make --allow-file-access-from-files and --start-fullscreen
work properly. The only thing that comes close it
using a new, private profile directory, so we do
exactly that. CKLaunch, Chrome.cmd and Sahli/_run.cmd
now use the temp/chrome directory for Chrome's
profile settings.

To make this possible, CKLaunch has been extended to
refer the the bin directory using the '?' character.