CKLaunch bugfix: changed order of $,&,? expansion
The order in which the special characters in [Actions]
were expanded turned out to be problematic: Since '$'
was expanded before '&', files with ampersands in their
names could no longer be started, as their names were
clobbered by the subsequent expansion of the '&' character.

The expansion order has now been changed to '&'->'$'->'?',
making use of the fact that question marks are practically
impossible to occur in file names, while there's nothing
special about dollars and ampersands, the latter being
in fact quite common (think "Artist A & Artist B - foo.mp3").