added DOSBox Staging as a standard package
Full integration is provided as usual. The CKLaunch default configuration
runs DOSBox Stating on .com files when pressing Alt+Enter. This keyboard
combination previously started DOSBox-X, which is now reachable with
Alt+Shift+Enter, but frankly, we can consider DOSBox-X as deprecated
at this point, as DOSBox Staging does most of what DOSBox-X did, and
it's more straightforward to configure.