add Foxotron
Don't hold your breath - this just makes setup.ps1 install it;
Foxotron otherwise doesn't play nice with the rest of CompoKit,
as it insists on being run from the program directory instead
of the document's directory and displays a demo-like config
dialog instead of going straight to desktop-resolution
fullscreen when run with a filename as a parameter.

Nevertheless, the common 3D scene file extensions (.dae, .fbx,
.glb etc.) have been registered with it in cklaunch.ini,
in the hope that a future version might fix the issues.