pm_slide_explort: option to export a functioning HTML page instead of PNG previews
Uses the HTML previews and glues many of them together to a single document,
peppered with a bit of JavaScript for slide navigation. The resulting document
still needs resources from the original webserver, so it's not fully stand-alone!
It's rather meant as a backup for when Slidemeister acts up (which is exactly
what it did the night before Deadline 2024, hence this option ...)